requirements.txt
asn1crypto==0.24.0
awsebcli==3.7.4
backcall==0.1.0
beautifulsoup4==4.6.0
botocore==1.10.54
cement==2.4.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.3
cryptography==1.9
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.5
django-autocomplete-light==3.2.10
django-dajax==0.9.2
django-dajaxice==0.7
django-debug-toolbar==1.9.1
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.14
elasticsearch==6.2.0
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.0
jmespath==0.9.3
jwt==0.5.3
lxml==4.2.1
Markdown==2.6.11
mysqlclient==1.3.12
openpyxl==2.5.3
parso==0.2.1
pathspec==0.3.3
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psutil==5.4.6
psycopg2==2.7.5
py==1.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pytest==3.2.5
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.13
requests==2.6.2
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
texttable==0.8.8
traitlets==4.3.2
urllib3==1.22
virtualenv==16.0.0
wcwidth==0.1.7
websocket-client==0.48.0
whitenoise==3.3.1
xlrd==1.1.0
